requirements.txt
certifi==2017.4.17
chardet==3.0.4
Django==1.10.6
django-cors-middleware==1.3.1
django-markdown-deux==1.0.5
django-registration==2.2
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
idna==2.5
markdown2==2.3.3
pkg-resources==0.0.0
PyJWT==1.4.2
PyMySQL==0.7.9
requests==2.18.1
urllib3==1.21.1
